pyproject.toml
[tool.poetry]
name = "contact-book"
version = "0.1.0"
description = ""
authors = ["sebastianLedzianowski <136706736+sebastianLedzianowski@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0.post1"
sqlalchemy = "^2.0.25"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
email-validator = "^2.1.0.post1"
pydantic = "^2.6.0"
python-dotenv = "^1.0.1"
faker = "^22.6.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.7"
libgravatar = "^1.0.4"
bcrypt = "^4.1.2"
fastapi-mail = "^1.4.1"
requests = "^2.31.0"
packaging = "^23.2"
redis = "^5.0.1"
fastapi-limiter = "^0.1.6"
cloudinary = "^1.38.0"
pydantic-settings = "^2.1.0"
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
httpx = "^0.26.0"
pytest-asyncio = "^0.23.5"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
